requirements.txt
torch>=2.3.0
transformers>=4.57.0
PyYAML>=6.0
sentencepiece>=0.2.0
accelerate>=0.33.0
modal>=1.1.0
peft>=0.17.0
